Don't use magic :{version} token in package.xml
It would be nice to use it, since this token is automatically replaced
with the version being released by bloom. If we don't use it, we need to
adjust the version manually whenever we do a new release. However, using
it produces the following fatal error when running catkin_make_isolated
on a source checkout of the package:

    Package version ":{version}" does not follow version conventions